Be more careful about the setting of mFileName. Since load() is called recursively for loading included files, we have to be sure that mFileName is set as the *last* step of loading rather than as the first so that the originally named file is the one that gets associated with the jccl::Configuration instance. Furthermore, during recursion, we cannot read from mFileName because its value is invalid at that point. Instead, we need to make careful use of scoping provided by the call stack to keep track of which file name really refers to the parent file.
A better way of handling this may be to have a separate method used for
the recursive inclusion process. Making this change would require breaking
up load() so that its code can be reused more effectively. The need for a
separate method to do the recursive inclusion is debatable.

This fixes Trac Ticket #2.


git-svn-id: https://vrjuggler.googlecode.com/svn/juggler/trunk@19811 08b38cba-cd3b-11de-854e-f91c5b6e4272