Use the actual version numbers for the Juggler modules instead of relying on substitution at build time. I decided to do this because I do not feel like making build_windows.py more complicated than it already is. The issue is that the variable containing the module version (such as GADGET_VERSION) gets used in two different contexts on Windows:
   1. In the substitution of @GADGET_VERSION@ in the generation of the .fpc
      files.
   2. As an environment variable used to set the version component of the
      DLL name in the Visual C++ build.

If it weren't for #2, we could use the major.minor.micro form for these
substitutions in the same way that we do on non-Windows platforms.

That said, doing the version requirements this way for Flagpoll is not so
bad. Given the goal of having binary compatibility between 2.2.x patch
releases, these requirements settings are perfectly valid. It's just messy.


git-svn-id: https://vrjuggler.googlecode.com/svn/juggler/branches/2.2@20617 08b38cba-cd3b-11de-854e-f91c5b6e4272