Added snx::sonix::shutdownAPI() so that user-level code has an opportunity to control when the sound API is shut down. Previously, this only happend in the snx::sonix destructor, a function that is only invoked when the snx::sonix singleton is destroyed. Technically, user-level code could have used snx::sonix::changeAPI() to switch to the "Stub" sound implementation, but that does not exactly have the desired outcome in my opinion.
This change allows for the VR Juggler Sound Manager to shut down Sonix so that
crash-on-exit problems in the Sonix plug-ins can be fixed. The issues with
Audiere and OpenAL seemed to be tied to threading. More specifically, the
sound APIs are created and initialized in the kernel control loop thread, but
they were destroyed in the primordial thread at the time of application exit.
Now, the Sound Manager can shut Sonix down from within the kernel control loop
thread.

Bumped the version to 1.3.6.


git-svn-id: https://vrjuggler.googlecode.com/svn/juggler/trunk@20871 08b38cba-cd3b-11de-854e-f91c5b6e4272