This change encompasses several things. The primary additions are that of gadget::DigitalEventInterface, gadget::MouseMultiClickEventInterface, and gadget::DigitalMultiClickEventInterface. The first adds an event interface for digital devices. The second and third are for multi-click events (double-click, triple-click, and so on).
As part of this work, I discovered a bug in the event generator implementation
with respect to device types that support multiple input units--in other words,
all device types except gadget::KeyboardMouse. Specifically, events were being
generated for all collected samples rather than for the unit that was
associated with the device proxy used by the event interface. Some refactoring
had to be done to address this bug, although the user-facing API has not
changed at all.

Bumped the version to 2.1.6.


git-svn-id: https://vrjuggler.googlecode.com/svn/juggler/trunk@21429 08b38cba-cd3b-11de-854e-f91c5b6e4272