Attempt a brute force fix for the problems with this script that were introduced in either r21626 or r21636 (I have forgotten which). This revision represents a simple copy of build_windows.py r21649 with adjustments made to account for details specific to the trunk.
Note that references to the vc10 directory tree have been removed. This
version of the script includes auto-migration support for the unattended
build. I have not had time to try migrating the vc9 solution to vc10, and
trying to maintain multiple versions of the Visual C++ build is painful
enough as it is.


git-svn-id: https://vrjuggler.googlecode.com/svn/juggler/trunk@21652 08b38cba-cd3b-11de-854e-f91c5b6e4272