Replaced gadget::InputMixer<T1,T2> and the various instantiations of that type (previously found in gadget/Type/InputBaseTypes.h) with a more flexible approach. The new base type for device drivers is gadget::InputDevice<T>. Its template paremeter is either a single device type such as gadget::Digital, gadget::Analog, etc., or it is a combination of device types created by boost::mpl::inheirt<T1,T2,...,Tn>::type.
Unlike the old approach with gadget::InputMixer, the ordering of types given
to boost::mpl::inherit does not matter. gadget::InputDevice<T> is implemented
so that the virtual methods readObject(), writeObject(), and getInputTypeName()
are invoked in a preset order. For example, using the base type
gadget::InputDevice<boost::mpl::inherit<gadget::Digital,gadget::Analog>::type>
is the same as using the base type
gadget::InputDevice<boost::mpl::inherit<gadget::Analog,gadget::Digital>::type>.
This will make things less error prone for device driver authors, and it means
that there is no longer a need to add a typedef each time a new device
combination is needed.

There is, however, still a pending issue. Just as before, all the currently
required input base types have to be enumerated in
gadget::BaseTypeFactory::registerBaseDeviceTypes(). I hope to replace this
enumeration with a compile-time algorithm that will generate all the base
type combinations (the unique subsets of gadget::all_base_types) so that it
never has to be updated again. My concern is that such an algorithm may be
too costly for compilers to handle, but I won't know that until I figure out
how to make it work.

Ultimately, this change is the first step in trying to make it easier to add
new devices types to Gadgeteer. There is currently a fair amount of code
duplication coupled with several (undocumented?)  extra steps that probably
make the process far too prone to mistakes and omissions.

Bumed the version to 2.1.18.


git-svn-id: https://vrjuggler.googlecode.com/svn/juggler/trunk@21735 08b38cba-cd3b-11de-854e-f91c5b6e4272