Configure input devices on slave nodes
The virtual devices on slave nodes are never configured and as a
consequence the base input types (Analog, Digital, ...) never have any
configuration applied.
This makes a noticeable difference for Analog when the device does not
produce values that are already in the normalized [0,1] range. A device
producing raw values in [-1,1] (and appropriately configured min/max
values) does not behave correctly on slave nodes, because the slave is
using the default min/max values instead of those specified in the config.

The solution is to forward config elements from InputDevice to all
bases and in the RIMPlugin pass the ConfigElement for a device to the
virtual device when it is first created.