Add option to set name of the local cluster_node
Before the cluster_node config element that applies to the local
machine is determined by comparing the machine's hostname with
the host_name property on the config element.
That makes it very difficult to run a cluster on the local machine for
testing purposes. If mLocalNodeName is set (is non-empty) its value is
used to determine the config element that applies to the local machine,
otherwise the existing behavior using the hostname is used.